NXT-186: Clear url when leaving the web ui
... and set the web-ui url again when switching to the web-ui.
The reason is to reload the web-ui completely everytime it's opened in
order to get the (changed) application state (e.g. opened workflows
etc.).
Switching back and forth between the classic UI and web UI shouldn't be a
common usage pattern and, thus, a constant synchronization between the
two states seems to be an overkill.

NXT-186 (Add perspective switch AP → AP.next)