Fix svg loader in Jest
The svg loader always loaded the same stub which caused incorrect test results.
With this change, all svg imports in tests will be treated as inline imports.
This also has potential of errors, but the only risk is that we identify an inline import
with a non-inline import (which currently does not exist).