NXT-1065: disable watcher even if the tooltip was closed by any other component
This fixes the odd apperance bug of old tooltips for 750ms while waiting
for the new one.
The watcher on the tooltip value was not removed if the tooltip was
closed via the store (by setting it to null).
In this case the watcher outlived the state and created problem.

NXT-1065 (Tooltips block UI before they are shown)