Base64 encode events (string messages) before passing it to the FE
... via the equo-comm service. Because the equo-comm service fails if
the string, e.g., contains escaped quotes ('\"'). Because quotes are
escaped again in the process which renders the orginal quote (which is
supposed to be escaped) 'active' again.
The FE then decodes it again.
Also: Most logic from equo-comm.js removed because it's now directly
part of the CEF's window object (EquoCommService).