NXT-1257: Make sure to always access the store to check for a feature-flag state
Because the feature flags are available in the application-store only _after_
the feature-flags-plugin has been executed. Also enables the feature flags
to be changed at runtime (e.g. through eclipse preferences) in the
future which are then passed to the frontend via an
application-state/store update.

NXT-1257 (Display embedded dialogs and views only when debug flag is true)