NXT-1285: Don't base64-encode events
This used to be necessary as a workaround to make the equo-comm-service
work. However, it has been fixed and we don't need it anymore.
It's also the culprit of this ticket because unicode and base64 don't
play together nicely.

NXT-1285 (Some characters e.g. “§”, “ラーメン大好き”, or “أنا أحب رامين” are replaced and additional  added in the CEF after using them for e.g. the component name)