NXT-1385: OpenWorkflowBrowserFunction re-worked
... to not show the java-based prompt anymore to the select the workflow
to open but open the workflow specified via browser-function parameters
directly.
It either opens the workflow in both, the Classic UI and Modern UI
(if the Classic UI is 'active'). Or only in the Modern UI (if the user
never opened the Classic UI so far).

NXT-1385 (Backend: Open workflow)