NXT-1385: Proper URI creation when opening workflow
Only relevant when opening a workflow in both, the classic and modern
UI. Because in order to get an instance of the filestore of the workflow
to be opened, we need to create an URI first. And this creation
previously failed in case of spaces in the workflow path. However, the
other URI-constructor now takes care of escaping spaces accordingly.

NXT-1385 (Backend: Open workflow)