NXT-1386: SaveAndCloseWorkflowsEvent introduced
Which generates all the svgs from the workflows to save and close and
passes them to the new 'saveAndCloseWorkflows'-browser function (along
with some extra parameters which are just forwarded).
Also: Loading-overlay introduced to prevent the user from changing
anything while svgs are being generated.

NXT-1386 (Backend: closing and saving a workflow independent from the classic UI)