NXT-1386: SaveAndCloseWorkflowsBrowserFunction-implementation introduced
Called from the FE when the BE requests to save and close workflows via a
respective event. Only purpose is to ask the FE to generate and transfer
the svgs for the workflows that need to be saved.

Also: propery clean-up of the KnimeBrowserView etc.

NXT-1386 (Backend: closing and saving a workflow independent from the classic UI)