NXT-1430: Close all WorkflowEditors on shutdown
... when shutting down from within the modern UI.
In order to do that a refactoring was necessary:
The shutdown-state-transition is triggered early now _before_ any
windows (and WorkflowEditors) are being closed. Triggered by the new
'PreShutdown'-extension points. The 'PreShutdown'-hook now also takes
care of the 'saveState' and 'suspend' state transitions instead of
callig them via the ISaveablePart2-mechanisms.
And: memorize the perspective on shutdown in the preference store ->
mainly required to be able to properly set the theme on the next
start-up (the modern UI and the classic UI require different themes).

NXT-1430 (Prevent welcome page and 'classic' workflow projects from being opened when AP is started with modern UI)