NXT-1430: Programmatically set 'perspective' system property
... on startup. What value to set is determined from a preference store
value; which is set on shutdown.
This is mainly used in order to be able to set the right them for
whatever perspective we start the AP with. Because the modern UI and
the classic UI require different themes.

NXT-1430 (Prevent welcome page and 'classic' workflow projects from being opened when AP is started with modern UI)