'isClassicPerspectiveActive'-flag wrongly initialized
It was always initialized with 'false' which lead to wrong behavior if
the AP was started with the classic UI, shutdown, and then started again
(erroneously with the modern UI then - because 'isClassicPerspectiveActive'
was still false).
The flag is now initialized according to the perspective the AP was
started with.