NXT-1486: Change order of checks when url changes
Only check if a URI can be imported if the other checks (i.e. whether
it's the webapp urls, a middleware-url, etc.) have been done.
This prevents a superfluous warning message in the console because,
e.g., the actual webapp-URL can't be imported, of course.

NXT-1486 (Add nodes to canvas via URI-drop and -paste)