NXT-1523: Only initate life cycle transition on perspective switch if possible
... when switching from the web UI to the java UI.
Because when starting with a 4.7 workspace, it switches first to the web
UI and the immediately back to the java UI without initializing anything in
the web UI perspective. And in order carry out every necessary step
require to switch back to the java UI, triggering the life cycle
transition must not fail.

NXT-1523 (First start with a 4.7 workspace ends up in classic UI (with modern UI theme))