NXT-1554: Special handling of links in the node description
The links now call a browser function in order to open the respective
URL in the external browser. Because otherwise, CEF will intercept the
link-URL and try to import it as a node if the link-URL is a
Hub-node-Link.

NXT-1554 (Some links in node description don’t work e.g. row aggregator)