AP-20228: Prevent NodeLogger-initialization on (very) early startup
Otherwise the workspace selection prompt doesn't show up.
Was cause by calling a static method on OpenKnimeUrlAction too early on
(which initiliazed the NodeLogger).
The respective method is now called later in the start-up (and on
ui-perspective switches).

AP-20228 (Restarting from classic perspective creates a new workspace and switches to it)