NXT-1728: Suspend Modern UI when switching to Classic UI on startup
Special case handling that's necessary for the following reason:
When starting a 5.x AP with a 4.x workspace for the very first time, it starts up with the Modern UI
(by default) but switches back again to the Classic UI during start-up (for unknown reasons).
I.e. the Modern UI has been initialized at this point but was never loaded (web app loaded state).
Thus this check to make sure we still suspend the Modern UI properly - which frees-up memory and
allows for a proper shutdown later on.

NXT-1728 (Fix startup routine for existing workspaces)