NXT-1728: Respect system property for 'isClassicPerspectiveActive'-flag
I.e. the flag also has a 'null'-state in which case the state is derived
from the 'perspective'-system property. This has the advantage that the
'isClassicPerspectiveActive'-flag doesn't need to be kept in sync with
the 'perspective'-property in case the perspective is never switched
(because in that case the 'perspective'-property state is the source of
truth which gets saved on shutdown as the 'startWithWebUI'-preference).

NXT-1728 (Fix startup routine for existing workspaces)