NXT-2019:  'checkForUpdate' desktop API function added
Such that it can be initiated from the frontend and is not done on the
backend side anymore.
This way we make the that the frontend is in a state where it can
receive the UpdateAvailableEvent.

NXT-2019 (Update event not reliably received by the frontend on AP startup)

NXT-2019: tmp

NXT-2019 (Update event not reliably received by the frontend on AP startup)