NXT-2062: Provide origin, do not consider temp copies as opened workflows
This reverts NXT-1885 and additionally adds frontend changes with the
goal that temporary copy workflows (i.e. those opened from a remote
space)...
- ... do not have an open indicator even though a corresponding temp
copy is open
- ... open multiple temp copies on doubleclick in space explorer

The functionality that the space explorer navigates to the space item
corresponding to the temp copy workflow is intentionally preserved.

NXT-2062 (Space explorer should navigate to location of current workflow if the workflow is opened via the space explorer)