NXT-2179: Activate/deactivate ProjectWorkflowMap
It's used in the classic UI to track workflow projects (for the UI and
workflow services) and is tied to the classic UI's WorkflowEditor.
I.e. it must not be used in case the modern UI is loaded (and no classic
UI loaded).

Deactivated on start-up with modern UI and activated on perspective
switch to classic UI (from then on it remains activated).

NXT-2179 (Workflows already loaded via 'Call Workflow (Service)' aren't taken into account)