NXT-2549: Remove logic to save project if no SVG was received.
We assume to always receive an SVG from the FE when saving a project, since that should only happen when in
'hybrid mode'. If we still don't receive one, see `generateWorkflowPreview.ts`, we log a warning and notify
the user with a Java pop-up.

NXT-2549 (Always close all workflows when switching between modern and classic UI)