NXT-135: Use RepositoryItem of correct version to create LocationInfo and WorkflowContext
Previously, it would default to the current-state version. `LocationInfo` does not seem to contain any fields specific to a version, but the `RepositoryItem` instance that is used to construct a `LocationInfo` is in fact specific to a version (contrary to as the name would imply).

This did not cause a visible issue because it appears none of the version-specific fields are used for that purpose. It is still not correct, though, to use an instance corresponding to an irrelevant, unrelated version. For example, some properties/behaviours may be added/changed in the future.

(NXT-135 Restore Version)