NXT-135: Restore Version
Defer instantiation of Projects to `Init`

Also for MostRecentlyUsedProjects for consistency. In other words, AppStatePersistor now only loads stuff from file into java objects but does not modify the state of the WebUI application such as e.g. ProjectManager.

Could be simplified if we would deserialize AppState always on `Init` (i.e. also on switch from Classic UI). See todos.

---

NXT-135: Do not construct caches for fixed version workflows

The state/caches of `WorkflowMiddleware` are only used for functionality that does not apply to fixed versions (workflow commands, change events, etc). It is enough to only supply the workflow entity.

Still the FE/BE Gateway API contains a "snapshot" property used for coordinating change events. Since this does not apply to fixed version workflows, it has been made optional/nullable.

Not going through WorkflowMiddleware for fixed version workflows also removes the need to clean up the cached state on version dispose, meaning the `WorkflowManagerCache` service dependency is no longer needed.

---
Further unify/streamline how `Project` instances are created to reduce ambiguity/duplication between AppStatePersistor, OpenProject and SaveProjectCopy. In all cases, `Project`s are created and loaded based on an `Origin`.

Change Lazy#ifInitialized to Lazy#ifPresent to be consistent with Optional#ifPresent

Use Project#updateOrigin instead of creating a new instance via builder to make sure that other properties (such as configured cache) stays intact.

Remove Project#Builder#onDispose. The same calls are already being made by ProjectWfmCache#disposeWorkflowManager via CoreUtil#cancelAndCloseLoadedWorkflow.

Remove duplicate method LocalSpace#getUniqueProjectId

---

De-duplicate wfm loading logic; Make wfm loading progress reporting a DesktopAPI dependency.

This was motivated by considering the necessity of a specific WorkflowManagerLoader that provides only the current version and fails if another is requested.

Testing code can now instead use a separate builder method as suggested.

 In "production" code, we create and open Projects via two places: (1) DesktopAPI>OpenProject and (2) AppStatePersistor>deserialize. The latter case initially featured such a WorkflowManagerLoader providing only a current version.

This is not necessary because, when deserializing open Projects, there is no inherent reason why the Project should not be able to provide versions as well. The constraint only came from that AppStatePersistor>deserialize had its own (duplicated) logic to load a single (!) WorkflowManager instance and then use this loaded instance to create a Project. However, from the perspective of gateway code, the restored Project might just as well also provide versions. Whether versions can be provided is rather a property of the Space(Provider).

So, this commit instead adjusts LocalSpace#toLocalAbsolutePath to fail if a version other than current-state is requested because the Space(Provider) simply does not support that.

This then means that there is no difference between how a WorkflowManager is loaded and a Project is created between (1) OpenProject and (2) AppStatePersistor. In both cases, we can now configure a project with a `WorkflowManagerLoader` that specifies how a WorkflowManager instance is loaded from the `Origin` of the Project. This eliminates the code duplication we previously had between (1) and (2).

Regarding production code, the WorkflowManagerLoader could simply be implemented as previously by `DesktopAPUtil.consumerWithProgress(..., DesktopAPUtil.fetchAndLoadWorkflowWithTask(...))`. The outer method provides an `IProgressMonitor` to the inner method to which it reports progress or exceptions. Concretely, up to now, this shows SWT progress bars or modals.

The problem with this, though, is that this does not work in headless unit tests (or generally any other cases in which you do not want to use that kind of progress reporting). This also seems to have been the motivation why `OpenProject` was previously structured as

```
// API entry point
static void openProject(...) {
  DesktopAPUtil.consumerWithProgress(..., openProjectWithProgress(...)
}

// only called from #openProject and tests
static void openProjectWithProgress(IProgressMonitor, ...) {
  // actually load and open the project
}
```
such that `openProjectWithProgress` can be called from `OpenProjectTest` with a `NullProgressMonitor` argument.

Instead of exposing methods to make it possible to configure their dependencies, we can also use the "service dependency" (or DesktopAPI dependency) concept which we already employ in many places (and in that specific test, too). This commit adds a DesktopAPI dependency `ProgressReporter` that can be implemented by `DesktopAPUtil#consumerWithProgress` (which is essentially `busyCursorWhile` plus catching) providing SWT modals. This makes it possible to test the code without making amends and allows some simplifications. Further, this factors out the SWT/workbench dependency and makes it ready to replace it: The `ProgressReporter` can readily become a gateway service dependency and, instead of showing a SWT progress modal, send progress info to the frontend, which may display a modern progress indicator which can also be used in a browser context.  (there's still more SWT dependency deeper in the callstack in `LoadWorkflowRunnable` though, but it's a start).

---

Add ability to dispose of a given workflow wfm, and to invoke the same loader when it is requested again

... as e.g. the case when we dispatch "restore this version to the working area" to the Hub catalog service. The local state of the working area then becomes invalid. Other use-case is deletion of a version that is locally cached, or "discard" of the working area.

---

Move Origin#of(HubSpaceLocationInfo, ...) to OpenProject
it is very specific to its usage there

Simplify SaveProjectCopy

Move Project#of to SaveProjectCopy#createProjectForCopiedWfm because the method is specific to that use-case.

Further "simplify" the control flow a bit: Saving of workflow copy and creation of project instance now happens in separate methods.

NXT-135 (Restore Version)