NXT-3150: Log DesktopAPI exceptions as errors
This means they will appear in the log file even if the user has not set their log level preference to `debug`. This is useful for diagnosing issues that people report -- previously, only the frontend exception would be logged and this could not provide information about the error.

Intermediate step until we have proper error handling and reporting for DesktopAPI calls.

NXT-3150 (Explicitly send consola logs to the backend)