NXT-3648: Run entire 'uri import' in separate thread
Because the uri-import logic uses Display.syncExec to synchronously run
stuff in the display thread. However, that otherwise causes the AP to
freeze on windows because that's not allowed anymore since NXT-3446
(CEF multi-threading on windows) when called from LocationListeners and
WindowOpenListeners.

NXT-3648 (Node DnD from Hub to AP opens internal browser window)