NXT-3825: Update only wfm name and origin in Project on Save-As
All other properties will remain untouched. See Project#of.

Further, generalise how Origin is inferred from Workflow(Context) to reduce specialisation and duplication.

And rename methods to make it clearer what happens "saveWorkflowCopy" might be misleading
because in fact the same WorkflowManager instance is saved, but it's save location
(LocationInfo, WorkflowContext) is modified before.
In other words, Save-As is a state operation on a
WorkflowManager instance and does not produce a new copy/instance.

NXT-3825 (Save-As yields workflow that cannot be executed)