NXT-3816: Do not show "Reveal in Space Explorer", "View Version History" if Space is not accessible ("unknown project").
We would previously only check whether the corresponding provider is connected, and even connect ad-hoc. This is not enough, since there may be spaces that are not in the user's available space groups. Example: One may be connected to the Community Hub, but the project may come from a public space of another user. Previously, when opening a public "unknown" project, the items would have been offered but failed on interaction.

NXT-3816 (Open of version workflow via URL drop does not show proper state in panel)