NXT-3816: Fix loading of known projects
The issue was that the previously-loaded wfm might have been a non-current-state wfm but we put it in the place the current-state wfm usually is.

This change avoids the need for that construct altogether by not loading the wfm in advance.

NXT-3816 (Open of version workflow via URL drop does not show proper state in panel)