NXT-3735: Fix definition of deserialised Projects
SaveProjectCopy#saveCopyOf would obtain the wfm to save as such:
```
final var wfm = project.loadWorkflowManager()
```
This is a mis-use as the wfm should not be loaded, we merely want to access the already-loaded instance.

It is understandable, though, that this happened. A `DefaultProject` is initialised with an already-loaded instance, and `loadWorkflowManager` just returns that instance.

In our case, the Project is not a `DefaultProject` but one created by AppStatePersistor#deserializeProject. The given implementation of `loadWorkflowManager` actually loads a wfm from disk. So, in the Save-As procedure we would load and save-as the wfm state from disk instead of the already-loaded state (potentially containing changes).

What we actually want for that use-case is lazy-loading of wfms: E.g. during appstate deserialisation we create `Project` instances but do not yet load the wfm instance.

This is not an issue in 5.5 because the lazy loading and access to wfms of projects has already been clarified there.

NXT-3735 (Save-As may lose changes)