Attach & dispose syncer instance cleanly -- A syncer ultimately is specific to a WorkflowManager instance because it listens to events emitted by a wfm. Indeed, currently, a Session corresponds to a Project corresponds to a WorkflowManager. However, the WorkflowManager instance is not guaranteed to be present for the entire session, see e.g. WorkflowService#disposeVersion. So we need to react to the wfm being disposed. This could be implemented by a Project#onWfmDispose callback (as in previous commits). But then in the callback implementation one would have to identify the syncer instance to call `dispose` on, i.e. have to know the project/session ID and then lookup the syncer in WorkflowSyncerManager. Attaching the syncer directly to the wfm via WorkflowResourceCache means (a) We do not need to add listener support in Project/ProjectWfmCache, (b) we do not need to register the on-dispose procedure of the syncer, complicated by a syncer only being optionally present, (c) we do not have to implement an access facility (previously the WorkflowSyncerManager/WorkflowSynerProvider service dependency) but can use existing API.
Note that two things have to be considered: (1) the syncer has to be attached to the wfm instance already loaded in GatewayJobPool (2) the syncer has to be attached to any wfm dynamically loaded in the future. For separation of concerns, (2) should not affect the WorkflowManagerLoader implementation (as in previous commits). (1) has to be done from the session initialisation. For (2) I can currently think of no better way than to expose a Project#onWfmLoad callback (as in previous commits).

Disposing a syncer instance is necessary because the sync job has to be canceled. For instance, a debounce timer might still be running.

Avoid having to pass projectId to calls like LocalSaver#saveProject -- this is needed only so that the implementation can access the wfm instance. This is counterintuitive since the syncer is already created specific to a wfm instance. Further, the wfm access would have to happen via static/singleton access (as in previous commits) or injected.